// src/store/user/user.types.ts
export enum USER_ACTIONS_TYPES {
  SET_CURRENT_USER = "SET_CURRENT_USER",
  CHECK_USER_SESSION = "CHECK_USER_SESSION",
  GOOGLE_SIGN_IN_START = "GOOGLE_SIGN_IN_START",
  EMAIL_SIGN_IN_START = "EMAIL_SIGN_IN_START",
  EMAIL_SIGNUP_START = "EMAIL_SIGNUP_START",
  SIGNOUT_START = "SIGNOUT_START",
  SIGNOUT_SUCCESS = "SIGNOUT_SUCCESS",
  SIGNOUT_FAILED = "SIGNOUT_FAILED",
  SIGN_IN_SUCCESS = "SIGN_IN_SUCCESS",
  SIGN_IN_FAILED = "SIGN_IN_FAILED",
  LOADING_STATUS = "LOADING_STATUS",
}